refactor(dist): drop OpenAPI
Drop OpenAPI spec for the job stream endpoint. It's not really meant for
external machine consumers, so it's likely just overhead to have it
here. It's primarily meant for our own Java tests (which can reuse the
same models from the endpoint directly) and for human readability.
